feat: 회원가입 api를 auth 폴더로 이동
