+++ src/main/java/kr/tennispark/auth/presentation/controller/AuthController.java
@@ -7,6 +7,7 @@ import kr.tennispark.auth.presentation.dto.request.VerifyPhoneRequest;
 import kr.tennispark.auth.presentation.dto.response.VerifyPhoneResponse;
 import kr.tennispark.common.utils.ApiUtils;
 import kr.tennispark.common.utils.ApiUtils.ApiResult;
+import kr.tennispark.members.user.presentation.dto.request.RegisterMemberRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,6 +23,13 @@ public class AuthController {
 
     private final AuthService authService;
 
+    @PostMapping
+    public ResponseEntity<ApiResult<String>> registerMember(@Valid @RequestBody RegisterMemberRequest request) {
+        authService.registerMember(request);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(ApiUtils.success());
+    }
+
     @PostMapping("/auth/phones/code")
     public ResponseEntity<ApiResult<String>> sendAuthCode(@Valid @RequestBody SendAuthCodeRequest request) {
         authService.sendAuthCode(request.phoneNumber());
@@ -29,10 +37,10 @@ public class AuthController {
                 .body(ApiUtils.success());
     }
 
+    @PostMapping("/auth/phones/code/verify")
     public ResponseEntity<ApiResult<VerifyPhoneResponse>> verifyPhone(@Valid @RequestBody VerifyPhoneRequest request) {
         VerifyPhoneResponse response = authService.verifyPhone(request);
         return ResponseEntity.status(HttpStatus.OK)
                 .body(ApiUtils.success(response));
     }
-
 }
