--- src/main/java/kr/tennispark/members/user/presentation/controller/MemberController.java
@@ -1,29 +1,0 @@
-package kr.tennispark.members.user.presentation.controller;
-
-import jakarta.validation.Valid;
-import kr.tennispark.common.utils.ApiUtils;
-import kr.tennispark.common.utils.ApiUtils.ApiResult;
-import kr.tennispark.members.user.application.service.MemberService;
-import kr.tennispark.members.user.presentation.dto.request.RegisterMemberRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/api/member")
-public class MemberController {
-
-    private final MemberService memberService;
-
-    @PostMapping
-    public ResponseEntity<ApiResult<String>> registerMember(@Valid @RequestBody RegisterMemberRequest request) {
-        memberService.registerMember(request);
-        return ResponseEntity.status(HttpStatus.CREATED)
-                .body(ApiUtils.success());
-    }
-}
